requirements.txt
bittensor
bittensor-wallet
python-statemachine
colorama
google-cloud-secret-manager
